fix: move out to function of get_playlist_thumbnail_path
